[fix] Migration

+++ bothub/common/models.py
@@ -1527,7 +1527,7 @@ class RepositoryEvaluateResultScore(models.Model):
 
     recall = models.DecimalField(max_digits=3, decimal_places=2, null=True)
 
-    support = models.IntegerField(null=True)
+    support = models.FloatField(null=True)
 
     created_at = models.DateTimeField(_("created at"), auto_now_add=True)
